+++ components/src/core/empty-state/empty-state.tsx
@@ -87,7 +87,8 @@ export const EmptyState: React.FC<EmptyStateProps> = (props) => {
 
     const getColor = useCallback(
         (color: string | undefined): string => {
-            if (!color) return theme.colors.text;
+            // @ts-ignore
+            if (!color) return theme.colors.textSecondary || theme.colors.text;
             if (Object.keys(theme.colors).indexOf(color) >= 0)
                 return theme.colors[color as keyof ReactNativePaper.Theme['colors']];
             return color;